Structural changes to better support rapidly playing sounds
